Move toolbar components into queryEditor dir
