+++ client/src/queryEditor/ToolbarSpacer.tsx
